+++ remind/src/App.js
@@ -4,39 +4,36 @@ import {
   Switch,
   Route,
 } from "react-router-dom";
-// import { auth } from './firebase/config';
-// import { useState, useEffect } from 'react';
+import { auth } from './firebase/config';
+import { useState, useEffect } from 'react';
 import Login from './components/login';
 import Register from './components/register';
-import Notes from './components/notes'
+import WallNotes from './components/wallNotes'
 import {Banner} from './components/banner';
 
 function App() {
- /*  const [user, setUser] = useState({});
+  const [user, setUser] = useState({});
         useEffect(() => {
         auth.onAuthStateChanged(user => {
             if(user) {
               setUser({email: user.email})
-            } else {
-              setUser({null:''})
             }
         })
     }, [])
-    console.log(user) */
-   // { email ? console.log('muestra /home') : console.log('regresa a /signin')}
+    // console.log(user)
   return (
     <Router>
         <Switch>
           <Route path="/" exact>
             <Banner/>
-            <Login/>
+            <Login setUser={setUser}/>
           </Route>
             <Route path="/signin">
               <Banner/>
               <Register/>
             </Route>
             <Route path="/wallNotes">
-              <Notes/>
+              <WallNotes user={user}/>
             </Route>
         </Switch> 
     </Router>
