+++ remind/src/components/modal.js
@@ -2,33 +2,33 @@ import React, { useState } from "react";
 import { createNotes } from '../firebase/firestore';
 import "./styles/modal.css"
 
-function Modal ({ showModal, setShowModal }) {
- const [title, setTitle] = useState('')
- const [note, setNote] = useState('')
+function Modal ({ showModal, setShowModal, user }) {
+ const [title, setTitle] = useState('');
+ const [note, setNote] = useState('');
+ 
 
   const handleSubmit = (e) => {
     e.preventDefault();
-     createNotes(title, note)
+     createNotes(title, note, user)
      .then(() => {
       console.log('Ya esta en FireStore')
       setShowModal((visible) => !visible)
      })
   };
   return showModal ? (
-    <div className="modal">
-      <form className="form-createNote" >
+    <section className="modal">
+      <form className="form-createNote"onSubmit={handleSubmit} >
       <div className="content-btn-close">
         <button className="btn-close" onClick={() => setShowModal((visible) => !visible)}>x</button>
       </div>
         <input onChange={(e) => {setTitle(e.target.value)}}
         name="title-note" id="title-note" placeholder="Título"/>
         <textarea onChange={(e) => {setNote(e.target.value)}} rows="15"
-        name="text-note" id="text-note" placeholder="Estribe tu nota">
+        name="text-note" id="text-note" placeholder="Escribe tu nota">
         </textarea>
-        <button onClick={handleSubmit}
-        type="submit" className="btn-add-note">Guardar</button>
+        <button type="submit" className="btn-add-note">Guardar</button>
       </form>
-    </div>
+    </section>
   ) : null;
 };
