+++ remind/src/firebase/firestore.js
@@ -1,8 +1,9 @@
 import { db } from "./config";
 
-export const createNotes = (title, note) => db.collection('reminds').doc().set({
+export const createNotes = (title, note, user) => db.collection('reminds').doc().set({
     title,
     note,
+    user,
 });
 export const showNotes = (querySnapshot) => db.collection('reminds').onSnapshot(querySnapshot);
