+++ remind/src/test/App.test.js
@@ -1,17 +1,17 @@
-import { render, screen } from "@testing-library/react";
-import App from '../App'
+// import { render, screen } from "@testing-library/react";
+// import App from '../App'
 
-beforeEach(() => render(<App/>));
+// beforeEach(() => render(<App/>));
 
-test("Cuando renderiza el componente Login", () => {
+// test("Cuando renderiza el componente Login", () => {
         
-        const contentEmail = screen.getByPlaceholderText('Correo electronico')
-        const contentPass = screen.getByPlaceholderText('Contraseña')
-        const contentSubmit = screen.getByRole('button', {name:/ENTRAR/i})
+//         const contentEmail = screen.getByPlaceholderText('Correo electronico')
+//         const contentPass = screen.getByPlaceholderText('Contraseña')
+//         const contentSubmit = screen.getByRole('button', {name:/ENTRAR/i})
 
-        expect(contentEmail).toBeInTheDocument()
-        expect(contentPass).toBeInTheDocument()
-        expect(contentSubmit).toBeInTheDocument()
+//         expect(contentEmail).toBeInTheDocument()
+//         expect(contentPass).toBeInTheDocument()
+//         expect(contentSubmit).toBeInTheDocument()
   
-});
+// });
