+++ remind/src/test/components/login.test.js
@@ -2,11 +2,11 @@ import { render, screen } from "@testing-library/react";
 import Login from '../../components/login'
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
-beforeEach(() => render(<Login/>));
+//beforeEach(() => render(<Login/>));
 
 
-test("Cuando renderiza el componente Register", () => {
-        
+test("Cuando renderiza el componente Login", () => {
+        beforeEach(() => render(<Login/>));
         const contentEmail = screen.getByPlaceholderText('Correo')
         const contentPass = screen.getByPlaceholderText('Contraseña')
         expect(contentEmail).toBeInTheDocument()
